FDS modulation & DPCM byte counter fixes
This branch will serve for future hotfixes:

* Fixed an FDS modulation issue in the sound engine affecting usage of delay
* Fixed DPCM "Bytes Left" counter so it correctly displays the remaining bytes

+++ FamiStudio/Source/Player/NesApu.cs
@@ -473,7 +473,7 @@ namespace FamiStudio
             public fixed byte Ages[20];
 
             // Extra internal states.
-            public byte DpcmBytesLeft;
+            public ushort DpcmBytesLeft;
             public byte DpcmDac;
         }
